requirements.txt
fastapi==0.95.0
pydantic==1.10.7
python-dotenv==1.0.0
requests==2.28.2
uvicorn==0.21.1
